Refactor hustle catalog into modular helpers
