+++ src/game/hustles/knowledgeHustles.js
@@ -1,0 +1,76 @@
+import { formatDays, formatHours, formatMoney } from '../../core/helpers.js';
+import { getState } from '../../core/state.js';
+import { executeAction } from '../actions.js';
+import { checkDayEnd } from '../lifecycle.js';
+import {
+  KNOWLEDGE_TRACKS,
+  enrollInKnowledgeTrack,
+  getKnowledgeProgress
+} from '../requirements.js';
+import { describeTrackEducationBonuses } from '../educationEffects.js';
+
+export function createKnowledgeHustles() {
+  return Object.values(KNOWLEDGE_TRACKS).map(track => ({
+    id: `study-${track.id}`,
+    studyTrackId: track.id,
+    name: track.name,
+    tag: { label: 'Study', type: 'study' },
+    description: track.description,
+    details: [
+      () => `🎓 Tuition: <strong>$${formatMoney(track.tuition)}</strong>`,
+      () => `⏳ Study Load: <strong>${formatHours(track.hoursPerDay)} / day for ${formatDays(track.days)}</strong>`,
+      () => {
+        const progress = getKnowledgeProgress(track.id);
+        if (progress.completed) {
+          return '✅ Status: <strong>Complete</strong>';
+        }
+        if (progress.enrolled) {
+          const remaining = Math.max(0, track.days - progress.daysCompleted);
+          return `📚 Status: <strong>${remaining} day${remaining === 1 ? '' : 's'} remaining</strong>`;
+        }
+        return '🚀 Status: <strong>Ready to enroll</strong>';
+      },
+      ...describeTrackEducationBonuses(track.id)
+    ],
+    action: {
+      id: `enroll-${track.id}`,
+      timeCost: 0,
+      moneyCost: Number(track.tuition) || 0,
+      label: () => {
+        const progress = getKnowledgeProgress(track.id);
+        if (progress.completed) return 'Course Complete';
+        if (progress.enrolled) {
+          const remaining = Math.max(0, track.days - progress.daysCompleted);
+          return remaining === 0 ? 'Graduation Pending' : `${remaining} day${remaining === 1 ? '' : 's'} remaining`;
+        }
+        const tuition = Number(track.tuition) || 0;
+        return tuition > 0 ? `Enroll for $${formatMoney(tuition)}` : 'Enroll Now';
+      },
+      className: 'secondary',
+      disabled: () => {
+        const state = getState();
+        const progress = getKnowledgeProgress(track.id);
+        if (progress.completed || progress.enrolled) return true;
+        const tuition = Number(track.tuition) || 0;
+        return tuition > 0 && state.money < tuition;
+      },
+      onClick: () => {
+        executeAction(() => {
+          const progress = getKnowledgeProgress(track.id);
+          if (progress.completed || progress.enrolled) return;
+          enrollInKnowledgeTrack(track.id);
+        });
+        checkDayEnd();
+      }
+    },
+    cardState: (_state, card) => {
+      if (!card) return;
+      const progress = getKnowledgeProgress(track.id);
+      card.classList.toggle('completed', progress.completed);
+      const inProgress = progress.enrolled && !progress.completed;
+      card.dataset.inProgress = inProgress ? 'true' : 'false';
+      card.dataset.studiedToday = progress.studiedToday ? 'true' : 'false';
+      card.dataset.enrolled = progress.enrolled ? 'true' : 'false';
+    }
+  }));
+}
